chore(admin-user-management): rename
user-management -> patient-management
